[expo] Apply code changes to universal module
