+++ android/src/main/java/expo/modules/blurview/BlurViewPackage.java
@@ -1,0 +1,16 @@
+package expo.modules.blurview;
+
+import android.content.Context;
+
+import java.util.Collections;
+import java.util.List;
+
+import expo.core.BasePackage;
+import expo.core.ViewManager;
+
+public class BlurViewPackage extends BasePackage {
+  @Override
+  public List<ViewManager> createViewManagers(Context context) {
+    return Collections.singletonList((ViewManager) new BlurViewManager());
+  }
+}
