--- ios/BlurViewManager.h
@@ -1,5 +1,0 @@
-#import <React/RCTViewManager.h>
-
-@interface BlurViewManager : RCTViewManager
-
-@end
