+++ ios/EXBlur.podspec
@@ -1,0 +1,23 @@
+require 'json'
+
+package = JSON.parse(File.read(File.join(__dir__, '..', 'package.json')))
+
+Pod::Spec.new do |s|
+  s.name           = 'EXBlur'
+  s.version        = package['version']
+  s.summary        = package['description']
+  s.description    = package['description']
+  s.license        = package['license']
+  s.author         = package['author']
+  s.homepage       = package['homepage']
+  s.platform       = :ios, '10.0'
+  s.source         = { git: 'https://github.com/expo/expo.git' }
+  s.source_files   = 'EXBlur/**/*.{h,m}'
+  s.preserve_paths = 'EXBlur/**/*.{h,m}'
+  s.requires_arc   = true
+
+  s.dependency 'EXCore'
+
+end
+
+
