+++ ios/EXBlur/EXBlurViewManager.h
@@ -1,0 +1,6 @@
+#import <EXCore/EXViewManager.h>
+#import <EXCore/EXModuleRegistryConsumer.h>
+
+@interface EXBlurViewManager : EXViewManager <EXModuleRegistryConsumer>
+
+@end
