+++ ios/EXBlur/EXVibrancyView.h
@@ -1,0 +1,4 @@
+#import <EXBlur/EXBlurView.h>
+
+@interface EXVibrancyView : EXBlurView
+@end
