+++ ios/EXBlur/EXVibrancyViewManager.h
@@ -1,0 +1,6 @@
+#import <EXCore/EXViewManager.h>
+#import <EXCore/EXModuleRegistryConsumer.h>
+
+@interface EXVibrancyViewManager : EXViewManager <EXModuleRegistryConsumer>
+
+@end
