--- ios/VibrancyView.h
@@ -1,4 +1,0 @@
-#import "BlurView.h"
-
-@interface VibrancyView : BlurView
-@end
