--- ios/VibrancyViewManager.h
@@ -1,5 +1,0 @@
-#import <React/RCTViewManager.h>
-
-@interface VibrancyViewManager : RCTViewManager
-
-@end
