+++ src/BlurView.ios.js
@@ -1,23 +1,21 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { View, requireNativeComponent, ViewPropTypes } from 'react-native';
+import { View, ViewPropTypes, findNodeHandle } from 'react-native';
+import { NativeModulesProxy, requireNativeViewManager } from 'expo-core';
 
 class BlurView extends Component {
   setNativeProps = nativeProps => {
     if (this._root) {
-      this._root.setNativeProps(nativeProps);
+      NativeModulesProxy.BlurViewManager.updateProps(nativeProps, findNodeHandle(this._root));
     }
-  }
+  };
 
   render() {
     return (
       <NativeBlurView
-        ref={e => this._root = e}
+        ref={e => (this._root = e)}
         {...this.props}
-        style={[
-          { backgroundColor: 'transparent' },
-          this.props.style,
-        ]}
+        style={[{ backgroundColor: 'transparent' }, this.props.style]}
       />
     );
   }
@@ -25,14 +23,7 @@ class BlurView extends Component {
 
 BlurView.propTypes = {
   ...(ViewPropTypes || View.propTypes),
-  blurType: PropTypes.oneOf([
-    'dark',
-    'light',
-    'xlight',
-    'prominent',
-    'regular',
-    'extraDark',
-  ]),
+  blurType: PropTypes.oneOf(['dark', 'light', 'xlight', 'prominent', 'regular', 'extraDark']),
   blurAmount: PropTypes.number,
 };
 
@@ -41,6 +32,6 @@ BlurView.defaultProps = {
   blurAmount: 10,
 };
 
-const NativeBlurView = requireNativeComponent('BlurView', BlurView);
+const NativeBlurView = requireNativeViewManager('BlurView');
 
 module.exports = BlurView;
