+++ src/VibrancyView.android.js
@@ -2,8 +2,9 @@ import React from 'react';
 
 class VibrancyView extends React.Component {
   render() {
-    console.error("VibrancyView is not implemented on Android");
+    console.error('VibrancyView is not implemented on Android');
+    return null;
   }
 }
 
-module.exports = VibrancyView
+module.exports = VibrancyView;
